+++ src/main/java/POT/DuoBloom/point/pointTransaction/entity/TransactionType.java
@@ -1,4 +1,4 @@
-package POT.DuoBloom.point.pointTransaction;
+package POT.DuoBloom.point.pointTransaction.entity;
 
 public enum TransactionType {
     CHALLENGE_REWARD,
