+++ src/main/java/POT/DuoBloom/point/pointTransaction/repository/PointTransactionRepository.java
@@ -1,10 +1,10 @@
-package POT.DuoBloom.point.pointTransaction;
+package POT.DuoBloom.point.pointTransaction.repository;
 
+import POT.DuoBloom.point.pointTransaction.entity.PointTransaction;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
-import java.util.Optional;
 
 @Repository
 public interface PointTransactionRepository extends JpaRepository<PointTransaction, Long> {
