+++ src/main/java/POT/DuoBloom/question/service/QuestionService.java
@@ -8,9 +8,9 @@ import POT.DuoBloom.question.mapper.AnswerMapper;
 import POT.DuoBloom.question.mapper.QuestionMapper;
 import POT.DuoBloom.question.repository.AnswerRepository;
 import POT.DuoBloom.question.repository.QuestionRepository;
-import POT.DuoBloom.point.pointTransaction.PointTransactionService;
-import POT.DuoBloom.point.pointTransaction.TransactionType;
-import POT.DuoBloom.user.UserRepository;
+import POT.DuoBloom.point.pointTransaction.service.PointTransactionService;
+import POT.DuoBloom.point.pointTransaction.entity.TransactionType;
+import POT.DuoBloom.user.repository.UserRepository;
 import POT.DuoBloom.user.entity.User;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
